Error handling
This is an example of proper error handling.

Your version was fine but wasn't scalable, which means you wouldnt be
able to scale the app easily going foreward without some code re-writing

e.g. you had the check for the zero error on the divide if statement,
this is cool but if we needed another bit of functionality to say  get a
percentage by dividing by 100 and multiplying by the percentage you
would need to rewrite the error check.

I've moved it into the divide function on the calculator class so that
every time you call the divide method you get the error check by default

I also changed your catch so that it spits out a more detailed message.
you're right in how you did it, you wouldnt want a user knowing what
went wrong but for this exercise we'll have it like this